N-gram tokenizers

These functions tokenize their inputs into different kinds of n-grams. The input can be a character vector of any length, or a list of character vectors where each character vector in the list has a length of 1. See details for an explanation of what each function does.

tokenize_ngrams(x, lowercase = TRUE, n = 3L, n_min = n,
  stopwords = character(), ngram_delim = " ", simplify = FALSE)

tokenize_skip_ngrams(x, lowercase = TRUE, n_min = 1, n = 3, k = 1,
  stopwords = character(), simplify = FALSE)

Arguments

x
A character vector or a list of character vectors to be tokenized into n-grams. If x is a character vector, it can be of any length, and each element will be tokenized separately. If x is a list of character vectors, each element of the list should have a length of 1.

lowercase
Should the tokens be made lower case?

n
The number of words in the n-gram. This must be an integer greater than or equal to 1.

n_min
This must be an integer greater than or equal to 1, and less than or equal to n.

stopwords
A character vector of stop words to be excluded from the n-grams.

ngram_delim
The separator between words in an n-gram.

simplify
FALSE by default so that a consistent value is returned regardless of length of input. If TRUE, then an input with a single element will return a character vector of tokens instead of a list.

k
For the skip n-gram tokenizer, the maximum skip distance between words. The function will compute all skip n-grams between 0 and k.

Value

A list of character vectors containing the tokens, with one element in the list for each element that was passed as input. If simplify = TRUE and only a single element was passed as input, then the output is a character vector of tokens.

Details

tokenize_ngrams:

Basic shingled n-grams. A contiguous subsequence of n words. This will compute shingled n-grams for every value of between n_min (which must be at least 1) and n.

tokenize_skip_ngrams:

Skip n-grams. A subsequence of n words which are at most a gap of k words between them. The skip n-grams will be calculated for all values from 0 to k.

These functions will strip all punctuation and normalize all whitespace to a single space character.

Examples

song <-  paste0("How many roads must a man walk down\n",
                "Before you call him a man?\n",
                "How many seas must a white dove sail\n",
                "Before she sleeps in the sand?\n",
                "\n",
                "How many times must the cannonballs fly\n",
                "Before they're forever banned?\n",
                "The answer, my friend, is blowin' in the wind.\n",
                "The answer is blowin' in the wind.\n")

tokenize_ngrams(song, n = 4)

#> [[1]]
#>  [1] "how many roads must"            "many roads must a"             
#>  [3] "roads must a man"               "must a man walk"               
#>  [5] "a man walk down"                "man walk down before"          
#>  [7] "walk down before you"           "down before you call"          
#>  [9] "before you call him"            "you call him a"                
#> [11] "call him a man"                 "him a man how"                 
#> [13] "a man how many"                 "man how many seas"             
#> [15] "how many seas must"             "many seas must a"              
#> [17] "seas must a white"              "must a white dove"             
#> [19] "a white dove sail"              "white dove sail before"        
#> [21] "dove sail before she"           "sail before she sleeps"        
#> [23] "before she sleeps in"           "she sleeps in the"             
#> [25] "sleeps in the sand"             "in the sand how"               
#> [27] "the sand how many"              "sand how many times"           
#> [29] "how many times must"            "many times must the"           
#> [31] "times must the cannonballs"     "must the cannonballs fly"      
#> [33] "the cannonballs fly before"     "cannonballs fly before they're"
#> [35] "fly before they're forever"     "before they're forever banned" 
#> [37] "they're forever banned the"     "forever banned the answer"     
#> [39] "banned the answer my"           "the answer my friend"          
#> [41] "answer my friend is"            "my friend is blowin"           
#> [43] "friend is blowin in"            "is blowin in the"              
#> [45] "blowin in the wind"             "in the wind the"               
#> [47] "the wind the answer"            "wind the answer is"            
#> [49] "the answer is blowin"           "answer is blowin in"           
#> [51] "is blowin in the"               "blowin in the wind"            
#> 

tokenize_ngrams(song, n = 4, n_min = 1)

#> [[1]]
#>   [1] "how"                            "how many"                      
#>   [3] "how many roads"                 "how many roads must"           
#>   [5] "many"                           "many roads"                    
#>   [7] "many roads must"                "many roads must a"             
#>   [9] "roads"                          "roads must"                    
#>  [11] "roads must a"                   "roads must a man"              
#>  [13] "must"                           "must a"                        
#>  [15] "must a man"                     "must a man walk"               
#>  [17] "a"                              "a man"                         
#>  [19] "a man walk"                     "a man walk down"               
#>  [21] "man"                            "man walk"                      
#>  [23] "man walk down"                  "man walk down before"          
#>  [25] "walk"                           "walk down"                     
#>  [27] "walk down before"               "walk down before you"          
#>  [29] "down"                           "down before"                   
#>  [31] "down before you"                "down before you call"          
#>  [33] "before"                         "before you"                    
#>  [35] "before you call"                "before you call him"           
#>  [37] "you"                            "you call"                      
#>  [39] "you call him"                   "you call him a"                
#>  [41] "call"                           "call him"                      
#>  [43] "call him a"                     "call him a man"                
#>  [45] "him"                            "him a"                         
#>  [47] "him a man"                      "him a man how"                 
#>  [49] "a"                              "a man"                         
#>  [51] "a man how"                      "a man how many"                
#>  [53] "man"                            "man how"                       
#>  [55] "man how many"                   "man how many seas"             
#>  [57] "how"                            "how many"                      
#>  [59] "how many seas"                  "how many seas must"            
#>  [61] "many"                           "many seas"                     
#>  [63] "many seas must"                 "many seas must a"              
#>  [65] "seas"                           "seas must"                     
#>  [67] "seas must a"                    "seas must a white"             
#>  [69] "must"                           "must a"                        
#>  [71] "must a white"                   "must a white dove"             
#>  [73] "a"                              "a white"                       
#>  [75] "a white dove"                   "a white dove sail"             
#>  [77] "white"                          "white dove"                    
#>  [79] "white dove sail"                "white dove sail before"        
#>  [81] "dove"                           "dove sail"                     
#>  [83] "dove sail before"               "dove sail before she"          
#>  [85] "sail"                           "sail before"                   
#>  [87] "sail before she"                "sail before she sleeps"        
#>  [89] "before"                         "before she"                    
#>  [91] "before she sleeps"              "before she sleeps in"          
#>  [93] "she"                            "she sleeps"                    
#>  [95] "she sleeps in"                  "she sleeps in the"             
#>  [97] "sleeps"                         "sleeps in"                     
#>  [99] "sleeps in the"                  "sleeps in the sand"            
#> [101] "in"                             "in the"                        
#> [103] "in the sand"                    "in the sand how"               
#> [105] "the"                            "the sand"                      
#> [107] "the sand how"                   "the sand how many"             
#> [109] "sand"                           "sand how"                      
#> [111] "sand how many"                  "sand how many times"           
#> [113] "how"                            "how many"                      
#> [115] "how many times"                 "how many times must"           
#> [117] "many"                           "many times"                    
#> [119] "many times must"                "many times must the"           
#> [121] "times"                          "times must"                    
#> [123] "times must the"                 "times must the cannonballs"    
#> [125] "must"                           "must the"                      
#> [127] "must the cannonballs"           "must the cannonballs fly"      
#> [129] "the"                            "the cannonballs"               
#> [131] "the cannonballs fly"            "the cannonballs fly before"    
#> [133] "cannonballs"                    "cannonballs fly"               
#> [135] "cannonballs fly before"         "cannonballs fly before they're"
#> [137] "fly"                            "fly before"                    
#> [139] "fly before they're"             "fly before they're forever"    
#> [141] "before"                         "before they're"                
#> [143] "before they're forever"         "before they're forever banned" 
#> [145] "they're"                        "they're forever"               
#> [147] "they're forever banned"         "they're forever banned the"    
#> [149] "forever"                        "forever banned"                
#> [151] "forever banned the"             "forever banned the answer"     
#> [153] "banned"                         "banned the"                    
#> [155] "banned the answer"              "banned the answer my"          
#> [157] "the"                            "the answer"                    
#> [159] "the answer my"                  "the answer my friend"          
#> [161] "answer"                         "answer my"                     
#> [163] "answer my friend"               "answer my friend is"           
#> [165] "my"                             "my friend"                     
#> [167] "my friend is"                   "my friend is blowin"           
#> [169] "friend"                         "friend is"                     
#> [171] "friend is blowin"               "friend is blowin in"           
#> [173] "is"                             "is blowin"                     
#> [175] "is blowin in"                   "is blowin in the"              
#> [177] "blowin"                         "blowin in"                     
#> [179] "blowin in the"                  "blowin in the wind"            
#> [181] "in"                             "in the"                        
#> [183] "in the wind"                    "in the wind the"               
#> [185] "the"                            "the wind"                      
#> [187] "the wind the"                   "the wind the answer"           
#> [189] "wind"                           "wind the"                      
#> [191] "wind the answer"                "wind the answer is"            
#> [193] "the"                            "the answer"                    
#> [195] "the answer is"                  "the answer is blowin"          
#> [197] "answer"                         "answer is"                     
#> [199] "answer is blowin"               "answer is blowin in"           
#> [201] "is"                             "is blowin"                     
#> [203] "is blowin in"                   "is blowin in the"              
#> [205] "blowin"                         "blowin in"                     
#> [207] "blowin in the"                  "blowin in the wind"            
#> [209] "in"                             "in the"                        
#> [211] "in the wind"                    "the"                           
#> [213] "the wind"                       "wind"                          
#> 

tokenize_skip_ngrams(song, n = 4, k = 2)

#> [[1]]
#>    [1] "how"                           "how many"                     
#>    [3] "how roads"                     "how must"                     
#>    [5] "how many roads"                "how many must"                
#>    [7] "how many a"                    "how roads must"               
#>    [9] "how roads a"                   "how roads man"                
#>   [11] "how must a"                    "how must man"                 
#>   [13] "how must walk"                 "how many roads must"          
#>   [15] "how many roads a"              "how many roads man"           
#>   [17] "how many must a"               "how many must man"            
#>   [19] "how many must walk"            "how many a man"               
#>   [21] "how many a walk"               "how many a down"              
#>   [23] "how roads must a"              "how roads must man"           
#>   [25] "how roads must walk"           "how roads a man"              
#>   [27] "how roads a walk"              "how roads a down"             
#>   [29] "how roads man walk"            "how roads man down"           
#>   [31] "how roads man before"          "how must a man"               
#>   [33] "how must a walk"               "how must a down"              
#>   [35] "how must man walk"             "how must man down"            
#>   [37] "how must man before"           "how must walk down"           
#>   [39] "how must walk before"          "how must walk you"            
#>   [41] "many"                          "many roads"                   
#>   [43] "many must"                     "many a"                       
#>   [45] "many roads must"               "many roads a"                 
#>   [47] "many roads man"                "many must a"                  
#>   [49] "many must man"                 "many must walk"               
#>   [51] "many a man"                    "many a walk"                  
#>   [53] "many a down"                   "many roads must a"            
#>   [55] "many roads must man"           "many roads must walk"         
#>   [57] "many roads a man"              "many roads a walk"            
#>   [59] "many roads a down"             "many roads man walk"          
#>   [61] "many roads man down"           "many roads man before"        
#>   [63] "many must a man"               "many must a walk"             
#>   [65] "many must a down"              "many must man walk"           
#>   [67] "many must man down"            "many must man before"         
#>   [69] "many must walk down"           "many must walk before"        
#>   [71] "many must walk you"            "many a man walk"              
#>   [73] "many a man down"               "many a man before"            
#>   [75] "many a walk down"              "many a walk before"           
#>   [77] "many a walk you"               "many a down before"           
#>   [79] "many a down you"               "many a down call"             
#>   [81] "roads"                         "roads must"                   
#>   [83] "roads a"                       "roads man"                    
#>   [85] "roads must a"                  "roads must man"               
#>   [87] "roads must walk"               "roads a man"                  
#>   [89] "roads a walk"                  "roads a down"                 
#>   [91] "roads man walk"                "roads man down"               
#>   [93] "roads man before"              "roads must a man"             
#>   [95] "roads must a walk"             "roads must a down"            
#>   [97] "roads must man walk"           "roads must man down"          
#>   [99] "roads must man before"         "roads must walk down"         
#>  [101] "roads must walk before"        "roads must walk you"          
#>  [103] "roads a man walk"              "roads a man down"             
#>  [105] "roads a man before"            "roads a walk down"            
#>  [107] "roads a walk before"           "roads a walk you"             
#>  [109] "roads a down before"           "roads a down you"             
#>  [111] "roads a down call"             "roads man walk down"          
#>  [113] "roads man walk before"         "roads man walk you"           
#>  [115] "roads man down before"         "roads man down you"           
#>  [117] "roads man down call"           "roads man before you"         
#>  [119] "roads man before call"         "roads man before him"         
#>  [121] "must"                          "must a"                       
#>  [123] "must man"                      "must walk"                    
#>  [125] "must a man"                    "must a walk"                  
#>  [127] "must a down"                   "must man walk"                
#>  [129] "must man down"                 "must man before"              
#>  [131] "must walk down"                "must walk before"             
#>  [133] "must walk you"                 "must a man walk"              
#>  [135] "must a man down"               "must a man before"            
#>  [137] "must a walk down"              "must a walk before"           
#>  [139] "must a walk you"               "must a down before"           
#>  [141] "must a down you"               "must a down call"             
#>  [143] "must man walk down"            "must man walk before"         
#>  [145] "must man walk you"             "must man down before"         
#>  [147] "must man down you"             "must man down call"           
#>  [149] "must man before you"           "must man before call"         
#>  [151] "must man before him"           "must walk down before"        
#>  [153] "must walk down you"            "must walk down call"          
#>  [155] "must walk before you"          "must walk before call"        
#>  [157] "must walk before him"          "must walk you call"           
#>  [159] "must walk you him"             "must walk you a"              
#>  [161] "a"                             "a man"                        
#>  [163] "a walk"                        "a down"                       
#>  [165] "a man walk"                    "a man down"                   
#>  [167] "a man before"                  "a walk down"                  
#>  [169] "a walk before"                 "a walk you"                   
#>  [171] "a down before"                 "a down you"                   
#>  [173] "a down call"                   "a man walk down"              
#>  [175] "a man walk before"             "a man walk you"               
#>  [177] "a man down before"             "a man down you"               
#>  [179] "a man down call"               "a man before you"             
#>  [181] "a man before call"             "a man before him"             
#>  [183] "a walk down before"            "a walk down you"              
#>  [185] "a walk down call"              "a walk before you"            
#>  [187] "a walk before call"            "a walk before him"            
#>  [189] "a walk you call"               "a walk you him"               
#>  [191] "a walk you a"                  "a down before you"            
#>  [193] "a down before call"            "a down before him"            
#>  [195] "a down you call"               "a down you him"               
#>  [197] "a down you a"                  "a down call him"              
#>  [199] "a down call a"                 "a down call man"              
#>  [201] "man"                           "man walk"                     
#>  [203] "man down"                      "man before"                   
#>  [205] "man walk down"                 "man walk before"              
#>  [207] "man walk you"                  "man down before"              
#>  [209] "man down you"                  "man down call"                
#>  [211] "man before you"                "man before call"              
#>  [213] "man before him"                "man walk down before"         
#>  [215] "man walk down you"             "man walk down call"           
#>  [217] "man walk before you"           "man walk before call"         
#>  [219] "man walk before him"           "man walk you call"            
#>  [221] "man walk you him"              "man walk you a"               
#>  [223] "man down before you"           "man down before call"         
#>  [225] "man down before him"           "man down you call"            
#>  [227] "man down you him"              "man down you a"               
#>  [229] "man down call him"             "man down call a"              
#>  [231] "man down call man"             "man before you call"          
#>  [233] "man before you him"            "man before you a"             
#>  [235] "man before call him"           "man before call a"            
#>  [237] "man before call man"           "man before him a"             
#>  [239] "man before him man"            "man before him how"           
#>  [241] "walk"                          "walk down"                    
#>  [243] "walk before"                   "walk you"                     
#>  [245] "walk down before"              "walk down you"                
#>  [247] "walk down call"                "walk before you"              
#>  [249] "walk before call"              "walk before him"              
#>  [251] "walk you call"                 "walk you him"                 
#>  [253] "walk you a"                    "walk down before you"         
#>  [255] "walk down before call"         "walk down before him"         
#>  [257] "walk down you call"            "walk down you him"            
#>  [259] "walk down you a"               "walk down call him"           
#>  [261] "walk down call a"              "walk down call man"           
#>  [263] "walk before you call"          "walk before you him"          
#>  [265] "walk before you a"             "walk before call him"         
#>  [267] "walk before call a"            "walk before call man"         
#>  [269] "walk before him a"             "walk before him man"          
#>  [271] "walk before him how"           "walk you call him"            
#>  [273] "walk you call a"               "walk you call man"            
#>  [275] "walk you him a"                "walk you him man"             
#>  [277] "walk you him how"              "walk you a man"               
#>  [279] "walk you a how"                "walk you a many"              
#>  [281] "down"                          "down before"                  
#>  [283] "down you"                      "down call"                    
#>  [285] "down before you"               "down before call"             
#>  [287] "down before him"               "down you call"                
#>  [289] "down you him"                  "down you a"                   
#>  [291] "down call him"                 "down call a"                  
#>  [293] "down call man"                 "down before you call"         
#>  [295] "down before you him"           "down before you a"            
#>  [297] "down before call him"          "down before call a"           
#>  [299] "down before call man"          "down before him a"            
#>  [301] "down before him man"           "down before him how"          
#>  [303] "down you call him"             "down you call a"              
#>  [305] "down you call man"             "down you him a"               
#>  [307] "down you him man"              "down you him how"             
#>  [309] "down you a man"                "down you a how"               
#>  [311] "down you a many"               "down call him a"              
#>  [313] "down call him man"             "down call him how"            
#>  [315] "down call a man"               "down call a how"              
#>  [317] "down call a many"              "down call man how"            
#>  [319] "down call man many"            "down call man seas"           
#>  [321] "before"                        "before you"                   
#>  [323] "before call"                   "before him"                   
#>  [325] "before you call"               "before you him"               
#>  [327] "before you a"                  "before call him"              
#>  [329] "before call a"                 "before call man"              
#>  [331] "before him a"                  "before him man"               
#>  [333] "before him how"                "before you call him"          
#>  [335] "before you call a"             "before you call man"          
#>  [337] "before you him a"              "before you him man"           
#>  [339] "before you him how"            "before you a man"             
#>  [341] "before you a how"              "before you a many"            
#>  [343] "before call him a"             "before call him man"          
#>  [345] "before call him how"           "before call a man"            
#>  [347] "before call a how"             "before call a many"           
#>  [349] "before call man how"           "before call man many"         
#>  [351] "before call man seas"          "before him a man"             
#>  [353] "before him a how"              "before him a many"            
#>  [355] "before him man how"            "before him man many"          
#>  [357] "before him man seas"           "before him how many"          
#>  [359] "before him how seas"           "before him how must"          
#>  [361] "you"                           "you call"                     
#>  [363] "you him"                       "you a"                        
#>  [365] "you call him"                  "you call a"                   
#>  [367] "you call man"                  "you him a"                    
#>  [369] "you him man"                   "you him how"                  
#>  [371] "you a man"                     "you a how"                    
#>  [373] "you a many"                    "you call him a"               
#>  [375] "you call him man"              "you call him how"             
#>  [377] "you call a man"                "you call a how"               
#>  [379] "you call a many"               "you call man how"             
#>  [381] "you call man many"             "you call man seas"            
#>  [383] "you him a man"                 "you him a how"                
#>  [385] "you him a many"                "you him man how"              
#>  [387] "you him man many"              "you him man seas"             
#>  [389] "you him how many"              "you him how seas"             
#>  [391] "you him how must"              "you a man how"                
#>  [393] "you a man many"                "you a man seas"               
#>  [395] "you a how many"                "you a how seas"               
#>  [397] "you a how must"                "you a many seas"              
#>  [399] "you a many must"               "you a many a"                 
#>  [401] "call"                          "call him"                     
#>  [403] "call a"                        "call man"                     
#>  [405] "call him a"                    "call him man"                 
#>  [407] "call him how"                  "call a man"                   
#>  [409] "call a how"                    "call a many"                  
#>  [411] "call man how"                  "call man many"                
#>  [413] "call man seas"                 "call him a man"               
#>  [415] "call him a how"                "call him a many"              
#>  [417] "call him man how"              "call him man many"            
#>  [419] "call him man seas"             "call him how many"            
#>  [421] "call him how seas"             "call him how must"            
#>  [423] "call a man how"                "call a man many"              
#>  [425] "call a man seas"               "call a how many"              
#>  [427] "call a how seas"               "call a how must"              
#>  [429] "call a many seas"              "call a many must"             
#>  [431] "call a many a"                 "call man how many"            
#>  [433] "call man how seas"             "call man how must"            
#>  [435] "call man many seas"            "call man many must"           
#>  [437] "call man many a"               "call man seas must"           
#>  [439] "call man seas a"               "call man seas white"          
#>  [441] "him"                           "him a"                        
#>  [443] "him man"                       "him how"                      
#>  [445] "him a man"                     "him a how"                    
#>  [447] "him a many"                    "him man how"                  
#>  [449] "him man many"                  "him man seas"                 
#>  [451] "him how many"                  "him how seas"                 
#>  [453] "him how must"                  "him a man how"                
#>  [455] "him a man many"                "him a man seas"               
#>  [457] "him a how many"                "him a how seas"               
#>  [459] "him a how must"                "him a many seas"              
#>  [461] "him a many must"               "him a many a"                 
#>  [463] "him man how many"              "him man how seas"             
#>  [465] "him man how must"              "him man many seas"            
#>  [467] "him man many must"             "him man many a"               
#>  [469] "him man seas must"             "him man seas a"               
#>  [471] "him man seas white"            "him how many seas"            
#>  [473] "him how many must"             "him how many a"               
#>  [475] "him how seas must"             "him how seas a"               
#>  [477] "him how seas white"            "him how must a"               
#>  [479] "him how must white"            "him how must dove"            
#>  [481] "a"                             "a man"                        
#>  [483] "a how"                         "a many"                       
#>  [485] "a man how"                     "a man many"                   
#>  [487] "a man seas"                    "a how many"                   
#>  [489] "a how seas"                    "a how must"                   
#>  [491] "a many seas"                   "a many must"                  
#>  [493] "a many a"                      "a man how many"               
#>  [495] "a man how seas"                "a man how must"               
#>  [497] "a man many seas"               "a man many must"              
#>  [499] "a man many a"                  "a man seas must"              
#>  [501] "a man seas a"                  "a man seas white"             
#>  [503] "a how many seas"               "a how many must"              
#>  [505] "a how many a"                  "a how seas must"              
#>  [507] "a how seas a"                  "a how seas white"             
#>  [509] "a how must a"                  "a how must white"             
#>  [511] "a how must dove"               "a many seas must"             
#>  [513] "a many seas a"                 "a many seas white"            
#>  [515] "a many must a"                 "a many must white"            
#>  [517] "a many must dove"              "a many a white"               
#>  [519] "a many a dove"                 "a many a sail"                
#>  [521] "man"                           "man how"                      
#>  [523] "man many"                      "man seas"                     
#>  [525] "man how many"                  "man how seas"                 
#>  [527] "man how must"                  "man many seas"                
#>  [529] "man many must"                 "man many a"                   
#>  [531] "man seas must"                 "man seas a"                   
#>  [533] "man seas white"                "man how many seas"            
#>  [535] "man how many must"             "man how many a"               
#>  [537] "man how seas must"             "man how seas a"               
#>  [539] "man how seas white"            "man how must a"               
#>  [541] "man how must white"            "man how must dove"            
#>  [543] "man many seas must"            "man many seas a"              
#>  [545] "man many seas white"           "man many must a"              
#>  [547] "man many must white"           "man many must dove"           
#>  [549] "man many a white"              "man many a dove"              
#>  [551] "man many a sail"               "man seas must a"              
#>  [553] "man seas must white"           "man seas must dove"           
#>  [555] "man seas a white"              "man seas a dove"              
#>  [557] "man seas a sail"               "man seas white dove"          
#>  [559] "man seas white sail"           "man seas white before"        
#>  [561] "how"                           "how many"                     
#>  [563] "how seas"                      "how must"                     
#>  [565] "how many seas"                 "how many must"                
#>  [567] "how many a"                    "how seas must"                
#>  [569] "how seas a"                    "how seas white"               
#>  [571] "how must a"                    "how must white"               
#>  [573] "how must dove"                 "how many seas must"           
#>  [575] "how many seas a"               "how many seas white"          
#>  [577] "how many must a"               "how many must white"          
#>  [579] "how many must dove"            "how many a white"             
#>  [581] "how many a dove"               "how many a sail"              
#>  [583] "how seas must a"               "how seas must white"          
#>  [585] "how seas must dove"            "how seas a white"             
#>  [587] "how seas a dove"               "how seas a sail"              
#>  [589] "how seas white dove"           "how seas white sail"          
#>  [591] "how seas white before"         "how must a white"             
#>  [593] "how must a dove"               "how must a sail"              
#>  [595] "how must white dove"           "how must white sail"          
#>  [597] "how must white before"         "how must dove sail"           
#>  [599] "how must dove before"          "how must dove she"            
#>  [601] "many"                          "many seas"                    
#>  [603] "many must"                     "many a"                       
#>  [605] "many seas must"                "many seas a"                  
#>  [607] "many seas white"               "many must a"                  
#>  [609] "many must white"               "many must dove"               
#>  [611] "many a white"                  "many a dove"                  
#>  [613] "many a sail"                   "many seas must a"             
#>  [615] "many seas must white"          "many seas must dove"          
#>  [617] "many seas a white"             "many seas a dove"             
#>  [619] "many seas a sail"              "many seas white dove"         
#>  [621] "many seas white sail"          "many seas white before"       
#>  [623] "many must a white"             "many must a dove"             
#>  [625] "many must a sail"              "many must white dove"         
#>  [627] "many must white sail"          "many must white before"       
#>  [629] "many must dove sail"           "many must dove before"        
#>  [631] "many must dove she"            "many a white dove"            
#>  [633] "many a white sail"             "many a white before"          
#>  [635] "many a dove sail"              "many a dove before"           
#>  [637] "many a dove she"               "many a sail before"           
#>  [639] "many a sail she"               "many a sail sleeps"           
#>  [641] "seas"                          "seas must"                    
#>  [643] "seas a"                        "seas white"                   
#>  [645] "seas must a"                   "seas must white"              
#>  [647] "seas must dove"                "seas a white"                 
#>  [649] "seas a dove"                   "seas a sail"                  
#>  [651] "seas white dove"               "seas white sail"              
#>  [653] "seas white before"             "seas must a white"            
#>  [655] "seas must a dove"              "seas must a sail"             
#>  [657] "seas must white dove"          "seas must white sail"         
#>  [659] "seas must white before"        "seas must dove sail"          
#>  [661] "seas must dove before"         "seas must dove she"           
#>  [663] "seas a white dove"             "seas a white sail"            
#>  [665] "seas a white before"           "seas a dove sail"             
#>  [667] "seas a dove before"            "seas a dove she"              
#>  [669] "seas a sail before"            "seas a sail she"              
#>  [671] "seas a sail sleeps"            "seas white dove sail"         
#>  [673] "seas white dove before"        "seas white dove she"          
#>  [675] "seas white sail before"        "seas white sail she"          
#>  [677] "seas white sail sleeps"        "seas white before she"        
#>  [679] "seas white before sleeps"      "seas white before in"         
#>  [681] "must"                          "must a"                       
#>  [683] "must white"                    "must dove"                    
#>  [685] "must a white"                  "must a dove"                  
#>  [687] "must a sail"                   "must white dove"              
#>  [689] "must white sail"               "must white before"            
#>  [691] "must dove sail"                "must dove before"             
#>  [693] "must dove she"                 "must a white dove"            
#>  [695] "must a white sail"             "must a white before"          
#>  [697] "must a dove sail"              "must a dove before"           
#>  [699] "must a dove she"               "must a sail before"           
#>  [701] "must a sail she"               "must a sail sleeps"           
#>  [703] "must white dove sail"          "must white dove before"       
#>  [705] "must white dove she"           "must white sail before"       
#>  [707] "must white sail she"           "must white sail sleeps"       
#>  [709] "must white before she"         "must white before sleeps"     
#>  [711] "must white before in"          "must dove sail before"        
#>  [713] "must dove sail she"            "must dove sail sleeps"        
#>  [715] "must dove before she"          "must dove before sleeps"      
#>  [717] "must dove before in"           "must dove she sleeps"         
#>  [719] "must dove she in"              "must dove she the"            
#>  [721] "a"                             "a white"                      
#>  [723] "a dove"                        "a sail"                       
#>  [725] "a white dove"                  "a white sail"                 
#>  [727] "a white before"                "a dove sail"                  
#>  [729] "a dove before"                 "a dove she"                   
#>  [731] "a sail before"                 "a sail she"                   
#>  [733] "a sail sleeps"                 "a white dove sail"            
#>  [735] "a white dove before"           "a white dove she"             
#>  [737] "a white sail before"           "a white sail she"             
#>  [739] "a white sail sleeps"           "a white before she"           
#>  [741] "a white before sleeps"         "a white before in"            
#>  [743] "a dove sail before"            "a dove sail she"              
#>  [745] "a dove sail sleeps"            "a dove before she"            
#>  [747] "a dove before sleeps"          "a dove before in"             
#>  [749] "a dove she sleeps"             "a dove she in"                
#>  [751] "a dove she the"                "a sail before she"            
#>  [753] "a sail before sleeps"          "a sail before in"             
#>  [755] "a sail she sleeps"             "a sail she in"                
#>  [757] "a sail she the"                "a sail sleeps in"             
#>  [759] "a sail sleeps the"             "a sail sleeps sand"           
#>  [761] "white"                         "white dove"                   
#>  [763] "white sail"                    "white before"                 
#>  [765] "white dove sail"               "white dove before"            
#>  [767] "white dove she"                "white sail before"            
#>  [769] "white sail she"                "white sail sleeps"            
#>  [771] "white before she"              "white before sleeps"          
#>  [773] "white before in"               "white dove sail before"       
#>  [775] "white dove sail she"           "white dove sail sleeps"       
#>  [777] "white dove before she"         "white dove before sleeps"     
#>  [779] "white dove before in"          "white dove she sleeps"        
#>  [781] "white dove she in"             "white dove she the"           
#>  [783] "white sail before she"         "white sail before sleeps"     
#>  [785] "white sail before in"          "white sail she sleeps"        
#>  [787] "white sail she in"             "white sail she the"           
#>  [789] "white sail sleeps in"          "white sail sleeps the"        
#>  [791] "white sail sleeps sand"        "white before she sleeps"      
#>  [793] "white before she in"           "white before she the"         
#>  [795] "white before sleeps in"        "white before sleeps the"      
#>  [797] "white before sleeps sand"      "white before in the"          
#>  [799] "white before in sand"          "white before in how"          
#>  [801] "dove"                          "dove sail"                    
#>  [803] "dove before"                   "dove she"                     
#>  [805] "dove sail before"              "dove sail she"                
#>  [807] "dove sail sleeps"              "dove before she"              
#>  [809] "dove before sleeps"            "dove before in"               
#>  [811] "dove she sleeps"               "dove she in"                  
#>  [813] "dove she the"                  "dove sail before she"         
#>  [815] "dove sail before sleeps"       "dove sail before in"          
#>  [817] "dove sail she sleeps"          "dove sail she in"             
#>  [819] "dove sail she the"             "dove sail sleeps in"          
#>  [821] "dove sail sleeps the"          "dove sail sleeps sand"        
#>  [823] "dove before she sleeps"        "dove before she in"           
#>  [825] "dove before she the"           "dove before sleeps in"        
#>  [827] "dove before sleeps the"        "dove before sleeps sand"      
#>  [829] "dove before in the"            "dove before in sand"          
#>  [831] "dove before in how"            "dove she sleeps in"           
#>  [833] "dove she sleeps the"           "dove she sleeps sand"         
#>  [835] "dove she in the"               "dove she in sand"             
#>  [837] "dove she in how"               "dove she the sand"            
#>  [839] "dove she the how"              "dove she the many"            
#>  [841] "sail"                          "sail before"                  
#>  [843] "sail she"                      "sail sleeps"                  
#>  [845] "sail before she"               "sail before sleeps"           
#>  [847] "sail before in"                "sail she sleeps"              
#>  [849] "sail she in"                   "sail she the"                 
#>  [851] "sail sleeps in"                "sail sleeps the"              
#>  [853] "sail sleeps sand"              "sail before she sleeps"       
#>  [855] "sail before she in"            "sail before she the"          
#>  [857] "sail before sleeps in"         "sail before sleeps the"       
#>  [859] "sail before sleeps sand"       "sail before in the"           
#>  [861] "sail before in sand"           "sail before in how"           
#>  [863] "sail she sleeps in"            "sail she sleeps the"          
#>  [865] "sail she sleeps sand"          "sail she in the"              
#>  [867] "sail she in sand"              "sail she in how"              
#>  [869] "sail she the sand"             "sail she the how"             
#>  [871] "sail she the many"             "sail sleeps in the"           
#>  [873] "sail sleeps in sand"           "sail sleeps in how"           
#>  [875] "sail sleeps the sand"          "sail sleeps the how"          
#>  [877] "sail sleeps the many"          "sail sleeps sand how"         
#>  [879] "sail sleeps sand many"         "sail sleeps sand times"       
#>  [881] "before"                        "before she"                   
#>  [883] "before sleeps"                 "before in"                    
#>  [885] "before she sleeps"             "before she in"                
#>  [887] "before she the"                "before sleeps in"             
#>  [889] "before sleeps the"             "before sleeps sand"           
#>  [891] "before in the"                 "before in sand"               
#>  [893] "before in how"                 "before she sleeps in"         
#>  [895] "before she sleeps the"         "before she sleeps sand"       
#>  [897] "before she in the"             "before she in sand"           
#>  [899] "before she in how"             "before she the sand"          
#>  [901] "before she the how"            "before she the many"          
#>  [903] "before sleeps in the"          "before sleeps in sand"        
#>  [905] "before sleeps in how"          "before sleeps the sand"       
#>  [907] "before sleeps the how"         "before sleeps the many"       
#>  [909] "before sleeps sand how"        "before sleeps sand many"      
#>  [911] "before sleeps sand times"      "before in the sand"           
#>  [913] "before in the how"             "before in the many"           
#>  [915] "before in sand how"            "before in sand many"          
#>  [917] "before in sand times"          "before in how many"           
#>  [919] "before in how times"           "before in how must"           
#>  [921] "she"                           "she sleeps"                   
#>  [923] "she in"                        "she the"                      
#>  [925] "she sleeps in"                 "she sleeps the"               
#>  [927] "she sleeps sand"               "she in the"                   
#>  [929] "she in sand"                   "she in how"                   
#>  [931] "she the sand"                  "she the how"                  
#>  [933] "she the many"                  "she sleeps in the"            
#>  [935] "she sleeps in sand"            "she sleeps in how"            
#>  [937] "she sleeps the sand"           "she sleeps the how"           
#>  [939] "she sleeps the many"           "she sleeps sand how"          
#>  [941] "she sleeps sand many"          "she sleeps sand times"        
#>  [943] "she in the sand"               "she in the how"               
#>  [945] "she in the many"               "she in sand how"              
#>  [947] "she in sand many"              "she in sand times"            
#>  [949] "she in how many"               "she in how times"             
#>  [951] "she in how must"               "she the sand how"             
#>  [953] "she the sand many"             "she the sand times"           
#>  [955] "she the how many"              "she the how times"            
#>  [957] "she the how must"              "she the many times"           
#>  [959] "she the many must"             "she the many the"             
#>  [961] "sleeps"                        "sleeps in"                    
#>  [963] "sleeps the"                    "sleeps sand"                  
#>  [965] "sleeps in the"                 "sleeps in sand"               
#>  [967] "sleeps in how"                 "sleeps the sand"              
#>  [969] "sleeps the how"                "sleeps the many"              
#>  [971] "sleeps sand how"               "sleeps sand many"             
#>  [973] "sleeps sand times"             "sleeps in the sand"           
#>  [975] "sleeps in the how"             "sleeps in the many"           
#>  [977] "sleeps in sand how"            "sleeps in sand many"          
#>  [979] "sleeps in sand times"          "sleeps in how many"           
#>  [981] "sleeps in how times"           "sleeps in how must"           
#>  [983] "sleeps the sand how"           "sleeps the sand many"         
#>  [985] "sleeps the sand times"         "sleeps the how many"          
#>  [987] "sleeps the how times"          "sleeps the how must"          
#>  [989] "sleeps the many times"         "sleeps the many must"         
#>  [991] "sleeps the many the"           "sleeps sand how many"         
#>  [993] "sleeps sand how times"         "sleeps sand how must"         
#>  [995] "sleeps sand many times"        "sleeps sand many must"        
#>  [997] "sleeps sand many the"          "sleeps sand times must"       
#>  [999] "sleeps sand times the"         "sleeps sand times cannonballs"
#>  [ reached getOption("max.print") -- omitted 996 entries ]
#>