Stopword lists

Retrieve lists of stopwords by language. The

stopwords(source = c("lucene", "jockers"), language = c("en", "da", "de",
  "el", "es", "fr", "it", "ru"))

Arguments

source
The name of the source of the stopwords.

language
The two-letter code for the name of the language.

References

The Lucene stopword lists are a subset of the stopword lists available in the Apache Lucene/Solr project, available under the Apache 2.0 license.

The Jockers stopword list is available from Matthew L. Jockers, "Expanded Stopwords List", http://www.matthewjockers.net/macroanalysisbook/expanded-stopwords-list/. According to Jockers, "The list includes the usual high frequency words ('the,' 'of,' 'an,' etc) but also several thousand personal names."

Examples

stopwords(language = "en")

#>  [1] "a"     "an"    "and"   "are"   "as"    "at"    "be"    "but"   "by"   
#> [10] "for"   "if"    "in"    "into"  "is"    "it"    "no"    "not"   "of"   
#> [19] "on"    "or"    "such"  "that"  "the"   "their" "then" 
#>  [ reached getOption("max.print") -- omitted 8 entries ]

stopwords(language = "de")

#>  [1] "aber"    "alle"    "allem"   "allen"   "aller"   "alles"   "als"    
#>  [8] "also"    "am"      "an"      "ander"   "andere"  "anderem" "anderen"
#> [15] "anderer" "anderes" "anderm"  "andern"  "anderr"  "anders"  "auch"   
#> [22] "auf"     "aus"     "bei"     "bin"    
#>  [ reached getOption("max.print") -- omitted 206 entries ]

head(stopwords(source = "jockers"), 20)

#>  [1] "0"       "1"       "2"       "3"       "4"       "5"       "6"      
#>  [8] "7"       "8"       "9"       "a"       "aaron"   "abbey"   "abbie"  
#> [15] "abdul"   "abe"     "across"  "abel"    "abigail" "about"