N-gram tokenizers

These functions tokenize their inputs into different kinds of n-grams. The input can be a character vector of any length, or a list of character vectors where each character vector in the list has a length of 1. See details for an explanation of what each function does.

tokenize_ngrams(x, lowercase = TRUE, n = 3L, n_min = n,
  stopwords = character(), ngram_delim = " ", simplify = FALSE)

tokenize_skip_ngrams(x, lowercase = TRUE, n_min = 1, n = 3, k = 1,
  stopwords = character(), simplify = FALSE)

Arguments

x
A character vector or a list of character vectors to be tokenized into n-grams. If x is a character vector, it can be of any length, and each element will be tokenized separately. If x is a list of character vectors, each element of the list should have a length of 1.

lowercase
Should the tokens be made lower case?

n
The number of words in the n-gram. This must be an integer greater than or equal to 1.

n_min
This must be an integer greater than or equal to 1, and less than or equal to n.

stopwords
A character vector of stop words to be excluded from the n-grams.

ngram_delim
The separator between words in an n-gram.

simplify
FALSE by default so that a consistent value is returned regardless of length of input. If TRUE, then an input with a single element will return a character vector of tokens instead of a list.

k
For the skip n-gram tokenizer, the maximum skip distance between words. The function will compute all skip n-grams between 0 and k.

Value

A list of character vectors containing the tokens, with one element in the list for each element that was passed as input. If simplify = TRUE and only a single element was passed as input, then the output is a character vector of tokens.

Details

tokenize_ngrams:

Basic shingled n-grams. A contiguous subsequence of n words. This will compute shingled n-grams for every value of between n_min (which must be at least 1) and n.

tokenize_skip_ngrams:

Skip n-grams. A subsequence of n words which are at most a gap of k words between them. The skip n-grams will be calculated for all values from 0 to k.

These functions will strip all punctuation and normalize all whitespace to a single space character.

Examples

song <-  paste0("How many roads must a man walk down\n",
                "Before you call him a man?\n",
                "How many seas must a white dove sail\n",
                "Before she sleeps in the sand?\n",
                "\n",
                "How many times must the cannonballs fly\n",
                "Before they're forever banned?\n",
                "The answer, my friend, is blowin' in the wind.\n",
                "The answer is blowin' in the wind.\n")

tokenize_ngrams(song, n = 4)

#> [[1]]
#>  [1] "how many roads must"    "many roads must a"      "roads must a man"      
#>  [4] "must a man walk"        "a man walk down"        "man walk down before"  
#>  [7] "walk down before you"   "down before you call"   "before you call him"   
#> [10] "you call him a"         "call him a man"         "him a man how"         
#> [13] "a man how many"         "man how many seas"      "how many seas must"    
#> [16] "many seas must a"       "seas must a white"      "must a white dove"     
#> [19] "a white dove sail"      "white dove sail before" "dove sail before she"  
#> [22] "sail before she sleeps" "before she sleeps in"   "she sleeps in the"     
#> [25] "sleeps in the sand"    
#>  [ reached getOption("max.print") -- omitted 27 entries ]
#> 

tokenize_ngrams(song, n = 4, n_min = 1)

#> [[1]]
#>  [1] "how"                  "how many"             "how many roads"      
#>  [4] "how many roads must"  "many"                 "many roads"          
#>  [7] "many roads must"      "many roads must a"    "roads"               
#> [10] "roads must"           "roads must a"         "roads must a man"    
#> [13] "must"                 "must a"               "must a man"          
#> [16] "must a man walk"      "a"                    "a man"               
#> [19] "a man walk"           "a man walk down"      "man"                 
#> [22] "man walk"             "man walk down"        "man walk down before"
#> [25] "walk"                
#>  [ reached getOption("max.print") -- omitted 189 entries ]
#> 

tokenize_skip_ngrams(song, n = 4, k = 2)

#> [[1]]
#>  [1] "how"                 "how many"            "how roads"          
#>  [4] "how must"            "how many roads"      "how many must"      
#>  [7] "how many a"          "how roads must"      "how roads a"        
#> [10] "how roads man"       "how must a"          "how must man"       
#> [13] "how must walk"       "how many roads must" "how many roads a"   
#> [16] "how many roads man"  "how many must a"     "how many must man"  
#> [19] "how many must walk"  "how many a man"      "how many a walk"    
#> [22] "how many a down"     "how roads must a"    "how roads must man" 
#> [25] "how roads must walk"
#>  [ reached getOption("max.print") -- omitted 1971 entries ]
#>